requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
httpx
cryptography
anthropic
openai
google-generativeai
gitpython
